EduLearn/requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
black==25.1.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-webpack-loader==3.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.17.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.1
